Removed: Ping no longer displays actually ping time (read description)
The actual ping time was removed so that all version-dependent code could be abolished from RoyalCommands. Unfortunately, getting the ping of a player requires accessing NMS (net.minecraft.server), which is version-dependent and not recommended by Bukkit. Until such a method of getting a player's ping is added to native Bukkit, actual ping times will be missing from this command.