Added: Safe NMS/CB internal support
This commit adds support for using CB/NMS code in a way that will not break servers or become incompatible in future releases. It is actually backwards compatible. I may make it so that NMS features are just disabled if this can't find support, allowing complete forwards compatibility.

On a related note, ping times are back in /ping.

Also, /setcharacteristic is added, rcmds.others.feed is in the plugin.yml, and teleports should remove momentum, as well.