Updated: Dynamic command registration system
In a nutshell, this allows for RoyalCommands to load commands at any time that the server is running. What it is used for now, however, is the disabling of commands. Before, disabling a command removed its functionality but left the command registered. Commands are no longer registered in the plugin.yml (!) as before, they are now under "reflectcommands" instead of "commands." Help had to be changed around a bit to accommodate that, as well. Disabling a command will now have the same effect as the command not existing at all. However, this has not been extensively tested (a few commands were sent in console, verifying that it worked - none as a player), so please be careful using this build.

Also, the way the default rules.txt, help.txt, and warps.yml are saved has been changed.

TL;DR: Disabling of commands works as expected. Untested.