Updated: Attempt to make opening OfflinePlayer inventories less exploitable
The way the system works is by checking the name of the inventory for needed information and then using that information to save the inventories. Now, due to inventories being able to be renamed, it was easy enough to exploit the saving process to overwrite inventories using renamed chests. This should make sure that only inventories not linked to blocks are checked, so only a malicious plugin could overwrite the inventory, which it could do by easier means.