Added: Vehicle teleports
This allows for any RoyalCommands teleport to find a vehicle that the player teleported is on and teleport /it/ instead, allowing for players in boats, on horses, etc. to be teleported. Animals, vehicles, and players that are being ridden all have a toggle in the config, as well as a main toggle for this feature altogether.