Fixed: Commands work on /pm update
This fixes an issue where using /pm update to update plugins would sometimes render all of that plugin's commands useless. To fix this, PluginManager unregisters all of the updated plugins commands before disabling it, which allows the new version to re-register them.

This also fixes the annoyance of having disabled plugins in the plugin list when using /pm update by removing the old plugin version if loading the new version was successful.

Loosely tested; use with caution.