Updated: Config refactor!
This update has completely redesigned the config.yml file. It now uses groups. This is not a final version of the configuration file, but it's a start. If you notice anything odd, please, please let me know. I have double-checked this, but it may not be perfect. If a value belongs in another group, or if a config value isn't changing anything (that means I didn't update it in the Config.java file), tell me. Basically, if you have any thoughts about this update, tell me.