Fixed: Remove blank line under item name with ProtocolLib
If you want to remove the blank line under the item name when spawned items are being tracked, just have ProtocolLib installed, and RoyalCommands will automagically make it disappear. This may cause stacking issues client-side, but the server should correct it.

A toggle will be added in a later commit.

This does not fix creative mode removing SpawnInfo, but it will when I receive word from another developer.