Updated: Try to fix issues with ProtocolLib
Ultimately, I haven't changed anything in terms of the issue. I did start using the official repository for AttributeStorage and the like (which required some reflection to get the custom features I added back). I'm not 100% sure the issue is my fault. ProtocolLib for 1.7.10 isn't fully supported, so maybe it'll be fixed in the future if I just leave it alone? I should talk to @aadnk and see if I'm doing something stupid wrong.