Added: New SubCommand system
Also, converted /pluginmanager to use it. This system will take a lot of the ugliness out of the commands that utilize subcommands. It has a registry system and an automatic help generator, so that I don't miss anything. It works really well in my testing, and tab-completion is also great.